+++ crates/biome_css_analyze/src/options.rs
@@ -23,7 +23,7 @@ pub type NoUnknownProperty =
     <lint::nursery::no_unknown_property::NoUnknownProperty as biome_analyze::Rule>::Options;
 pub type NoUnknownPseudoClass =
     <lint::nursery::no_unknown_pseudo_class::NoUnknownPseudoClass as biome_analyze::Rule>::Options;
-pub type NoUnknownSelectorPseudoElement = < lint :: nursery :: no_unknown_selector_pseudo_element :: NoUnknownSelectorPseudoElement as biome_analyze :: Rule > :: Options ;
+pub type NoUnknownPseudoElement = < lint :: nursery :: no_unknown_pseudo_element :: NoUnknownPseudoElement as biome_analyze :: Rule > :: Options ;
 pub type NoUnknownUnit =
     <lint::nursery::no_unknown_unit::NoUnknownUnit as biome_analyze::Rule>::Options;
 pub type NoUnmatchableAnbSelector = < lint :: nursery :: no_unmatchable_anb_selector :: NoUnmatchableAnbSelector as biome_analyze :: Rule > :: Options ;